+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v0.12.7] - 2022-09-08
+
+### Added
+- `--region` flag with `REGION` env option to the `zcli bucket s3` `create` and `delete` commands
+ 
 ## [v0.12.6] - 2022-09-01
 
 ### Fixed